Add Catch2 tests for Input_file_segy and Output_file_segy
Add "better" tests for Input_file_segy and Output_file_segy.
Parsing of SEG-Y files are based on a "test" SEG-Y file where the binary
and native representations of the file are generated at the same time.

A number of other facilities had tests ported to Catch2 as well, particularly
the number_encoding and character_encoding facilities.

Change Log:
* Add Catch2 tests for Input_file_segy and Output_file_segy
* Add Catch2 tests for character_encoding and number_encoding
* Add test SEG-Y file generator
* Fix incorrect assertion in IO_driver_distributed_vector::read_strided and
    write_strided
* Add exseis::test::Index_hash facility for generating random numbers that can
    be quickly re-generated during testing
* Improve development flags
* Rename /utils/ directory to /apps/ to be more inline with the Pitchfork Layout
* Make "segy" namespaces non-inline to make SEG-Y specific parts stand out more