Cleanup layout and structure
Updated some filenames and directory names to better reflect their typical
namespacing.

Removed the ExSeis and ExSeisPIOL classes, and replaced them with passing in
a Communicator or Communicator_mpi.
Removed the implicit construction of IO_driver_mpi in Input_file_segy and
Output_file_segy with just passing in IO_driver_mpi.
Generally improved the implementation of Dependency Injection in the project.

This necessitated a cascade of edits to allow Input_file_mpi to get its
communicator, logger, and file name from IO_driver_mpi, along with several other
instances of needing to pull a communicator from a class used for construction.
For example, the Communicator now has a method to return a Log instance.
This is intended to allow different parallel frameworks to specialize how
IO and communication with the outside world (e.g. printing to screen, reporting
errors). Some further work would be needed to enable Log to be specialized,
but the ground work is there.

Made `piol` and `utils` namespaces inline. All main classes should be accessible
using just `exseis::`. Removed instances of `using namespace exseis::piol` and
`using namespace exseis::utils` in favour of explicit `exseis::` on symbols.
With a shorter namespace invocation, this is now feasible.

Change Log:
* Rename include/exseisdat to include/exseis
* Restructure and rename files and folders in apps/
* Remove ExSeis and ExSeisPIOL classes
* Add various methods to enable better Dependency Injection
* Change macro prefixes from EXSEISDAT_ to EXSEIS_
* Set piol and utils namespaces to inline
* Remove usage of `using namespace exseis::piol` and `... exseis::utils` and
  use explicit `exseis::` in `test/` `apps/` and `examples/`